[POTENTIAL BREAK!] Updates
Updates to verify_credentials() to check if the credentials have
expired, and if so, reauth instead of attempting to refresh an expired
access token. Improves control/handling of OAuth errors.

Ensure that user information is returned on successful auth, rather
than just relying on the isReady() method to check after the call

setUser() now also required the access_token and handles storing in
session

unique token locations in session (unique to application) via APP_NAME
constant which MUST be defined at the top of the script